requirements.txt
Flask>=2.3.0,<3.0.0
Flask-SQLAlchemy>=3.0.0,<4.0.0
Flask-Login>=0.6.0,<1.0.0
Werkzeug>=2.3.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
Authlib>=1.2.0,<2.0.0
PyPDF2>=3.0.0,<4.0.0
python-docx>=0.8.0,<1.0.0
easyocr>=1.7.0,<2.0.0
Pillow>=10.0.0,<11.0.0
opencv-python>=4.8.0,<5.0.0
numpy>=1.24.0,<2.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
duckduckgo-search>=3.9.0,<4.0.0
google-generativeai>=0.3.0,<1.0.0
langdetect>=1.0.0,<2.0.0
waitress>=3.0.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
